# ENH: sparse.csgraph.yen: performance improvements