#### Reference issue
Closes [gh-20366.](https://github.com/scipy/scipy/issues/20366)

#### What does this implement?
1. Introduce a benchmark for Yen's algorithm, allowing measuring performance improvement of following commits.
2. Create a dedicated class to manage the container of found paths.
3. Introduce [Lawler's modification](https://en.wikipedia.org/wiki/Yen%27s_algorithm#Lawler's_modification) to Yen's algorithm.

#### Additional information
We use a **sorted vector** of varying size to store the found paths. This container allows:
1. Access to the minimal and maximal distances in the container in O(1).
2. Insertion in O(K).
3. Controlled size - this is an important performance factor because when the graphs are large- plenty of possible paths are found. Data structures that don't allow controlling size (e.g., `priority_queue`) tend to be huge, and managing them such data structures impacts performance significantly.

After trying multiple possible containers, this one gives the best performance/complexity trade-off balance.
Remember that managing this container is a fraction of the Yen's algorithm total run-time.

#### Performance comparison

Adding the managed sorted vector container for found paths does not noticeably impact performance.
```
python dev.py bench -t sparse_csgraph_yen.Yen --compare b36655bc0 --compare 83d3beb2c
...
(Sorted vector)
             --                      K                  
             ------ ------------------------------------
               n        10         100          300     
             ====== ========== ============ ============
               30    411±20μs   2.41±0.2ms    8.11±2ms  
              300    6.15±1ms    56.8±3ms     225±20ms  
              3000   541±40ms   4.18±0.2s    12.6±0.01s 
             ====== ========== ============ ============

...
(Benchmark over main)
              --                      K                  
              ------ ------------------------------------
                n        10         100          300     
              ====== ========== ============ ============
                30    396±30μs   2.61±0.2ms   9.33±0.3ms 
               300    6.30±1ms   52.6±10ms     183±10ms  
               3000   608±40ms   4.26±0.05s   11.4±0.2s  
              ====== ========== ============ ============
BENCHMARKS NOT SIGNIFICANTLY CHANGED.
```

However, Lawler's modification does improve performance significantly
```
python dev.py bench -t sparse_csgraph_yen.Yen --compare acc857e24 --compare b36655bc0
...
(Lawler's modification)
             --                       K                   
             ------ --------------------------------------
               n         10          100          300     
             ====== ============ ============ ============
               30     431±90μs    1.86±0.2ms   4.16±0.1ms 
              300    4.55±0.6ms    43.1±9ms     164±50ms  
              3000   457±100ms    3.62±0.3s    10.3±0.2s  
             ====== ============ ============ ============

...
(Sorted vector)
              --                       K                   
              ------ --------------------------------------
                n         10          100          300     
              ====== ============ ============ ============
                30     464±50μs    3.19±0.3ms   8.46±0.5ms 
               300    5.76±0.5ms    52.3±5ms     208±10ms  
               3000    593±40ms    4.87±0.07s    18.6±2s   
              ====== ============ ============ ============

| Change   | Before [b36655bc0] <sery/enh_yen~1>   | After [acc857e24 ] <sery/enh_yen>   |   Ratio | Benchmark (Parameter)                    |
|----------|--------------------------------------|-----------------------------------|---------|------------------------------------------|
| -        | 3.19±0.3ms                           | 1.86±0.2ms                        |    0.58 | sparse_csgraph_yen.Yen.time_yen(30, 100) |
| -        | 8.46±0.5ms                           | 4.16±0.1ms                        |    0.49 | sparse_csgraph_yen.Yen.time_yen(30, 300) |

```