@tsery-ns 2025-01-15
Hi, I need help to solve this one:
```
[32.56%] ··· sparse_csgraph_yen.Yen.time_yen                             failed
[32.56%] ··· ====== ======== ======== ========
             --                 K             
             ------ --------------------------
               n       10      100      300   
             ====== ======== ======== ========
               30    failed   failed   failed 
              300    failed   failed   failed 
              3000   failed   failed   failed 
             ====== ======== ======== ========
```
Is this timeout? should I reduce the test time?
Thank you.