@czgdp1807 2025-01-17
Benchmarking code,

```python
from scipy.sparse.csgraph import yen
import numpy as np, scipy, timeit, functools
from tabulate import tabulate

def scipy_yen(graph, source, sink, K):
    yen(graph, source, sink, K)

N = 500
p_m = u" \u00B1 "
mu = u" \u03BC "
sigma = u" \u03C3 "
results = []
densities = [0.1, 0.3, 0.5, 0.9]
K = [2, N//7, N//6]
sinks = [10, 100]
for density in densities:
    A = (scipy.sparse.rand(N, N, density, format='csr', rng=0)*100).astype(np.uint32)
    for k in K:
        for sink in sinks:
            try:
                t = timeit.Timer(functools.partial(scipy_yen, A, 0, sink, k))
                times = t.repeat(3, 1)
                result = [density]
                result.append('Yen')
                result.append(k)
                result.append(0)
                result.append(sink)
                result.append(str(np.mean(times).round(3)) + p_m + str(np.std(times).round(3)))

                results.append(result)
            except ValueError:
                pass

print(tabulate(results, headers=[
    'Density', 'Algorithm', 'K', 'Source', 'Sink',
    mu + p_m + sigma],
    tablefmt='github'), end="\n\n")
```

On `main` branch,

```zsh
|   Density | Algorithm   |   K |   Source |   Sink |  μ  ±  σ      |
|-----------|-------------|-----|----------|--------|---------------|
|       0.1 | Yen         |   2 |        0 |     10 | 0.001 ± 0.0   |
|       0.1 | Yen         |   2 |        0 |    100 | 0.001 ± 0.0   |
|       0.1 | Yen         |  71 |        0 |     10 | 0.158 ± 0.001 |
|       0.1 | Yen         |  71 |        0 |    100 | 0.166 ± 0.001 |
|       0.1 | Yen         |  83 |        0 |     10 | 0.195 ± 0.004 |
|       0.1 | Yen         |  83 |        0 |    100 | 0.192 ± 0.001 |
|       0.3 | Yen         |   2 |        0 |     10 | 0.013 ± 0.0   |
|       0.3 | Yen         |   2 |        0 |    100 | 0.006 ± 0.0   |
|       0.3 | Yen         |  71 |        0 |     10 | 1.306 ± 0.001 |
|       0.3 | Yen         |  71 |        0 |    100 | 1.133 ± 0.001 |
|       0.3 | Yen         |  83 |        0 |     10 | 1.504 ± 0.002 |
|       0.3 | Yen         |  83 |        0 |    100 | 1.196 ± 0.002 |
|       0.5 | Yen         |   2 |        0 |     10 | 0.007 ± 0.0   |
|       0.5 | Yen         |   2 |        0 |    100 | 0.009 ± 0.0   |
|       0.5 | Yen         |  71 |        0 |     10 | 4.203 ± 0.003 |
|       0.5 | Yen         |  71 |        0 |    100 | 4.306 ± 0.003 |
|       0.5 | Yen         |  83 |        0 |     10 | 4.846 ± 0.017 |
|       0.5 | Yen         |  83 |        0 |    100 | 4.998 ± 0.003 |
|       0.9 | Yen         |   2 |        0 |     10 | 0.005 ± 0.0   |
|       0.9 | Yen         |   2 |        0 |    100 | 0.007 ± 0.0   |
|       0.9 | Yen         |  71 |        0 |     10 | 7.997 ± 0.0   |
|       0.9 | Yen         |  71 |        0 |    100 | 7.993 ± 0.019 |
|       0.9 | Yen         |  83 |        0 |     10 | 9.79 ± 0.01   |
|       0.9 | Yen         |  83 |        0 |    100 | 9.953 ± 0.006 |
```

On this PR,

```zsh
(scipy-dev) 13:33:14:~/Quansight/scipy % python dev.py python ../debug.py
💻  ninja -C /Users/czgdp1807/Quansight/scipy/build -j10
ninja: Entering directory `/Users/czgdp1807/Quansight/scipy/build'
[2/338] Generating subprojects/highs/src/HConfig.h with a custom command
Build OK
💻  meson install -C build --only-changed --tags runtime,python-runtime,tests,devel
Installing, see meson-install.log...
Installation OK
zsh: abort      python dev.py python ../debug.py
```

Probably there is some bug here, because on `main` branch the same code works fine.